StautDef

Documentation

translate :: Map FuncId (FuncDef VarId) -> Id -> Id -> Name -> [ChanId] -> [VarId] -> ExitSort -> [StatId] -> [VarId] -> [Trans] -> StatId -> VEnv -> (ProcDef, BExpr)

combineParameters :: [VarId] -> VarId -> [VarId] -> [VarId]

combine parameters of stautdef, state identifier, and stautdef variables into single parameter list must be consistent with combineArguments